update / fix examples
