+++ examples/greeter/greeter.go
@@ -2,13 +2,17 @@ package main
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/hashicorp/go-hclog"
 	"github.com/treethought/roc"
 	"github.com/treethought/roc/proto"
 )
 
-var log = hclog.Default()
+var log = hclog.New(&hclog.LoggerOptions{
+	Color:  hclog.AutoColor,
+	Output: os.Stdout,
+})
 
 type MyEndpoint struct {
 	*roc.Accessor
@@ -22,55 +26,35 @@ func New() *MyEndpoint {
 
 // Source retrieves representation of resource
 func (e *MyEndpoint) Source(ctx *roc.RequestContext) interface{} {
-	log.Error("Executing source in greeter", "identifier", ctx.Request().Identifier)
-
-	log.Warn("Making subrequest", "target", "res://name")
-	log.Warn("sourcing http request")
-
-	// httpReq, err := ctx.Source(roc.NewIdentifier("httpRequest://params"), nil)
-	// if err != nil {
-	// 	log.Error("failed to get httpRequest://", "err", err)
-	// 	return &proto.ErrorMessage{Message: err.Error()}
-	// }
-
+	// get the argument value from the identifier provided as request argument
 	name, err := ctx.GetArgumentValue("name")
 	if err != nil {
-		log.Error("failed to get name argument", "err", err)
-		// return roc.NewRepresentation(&proto.ErrorMessage{Message: err.Error)})
+		log.Warn("failed to get name argument", "err", err)
+		return "no name?"
 	}
 
-	m := new(proto.String)
-	err = name.MarshalTo(m)
-	if err != nil {
-		log.Error("failed to marshal argument to string", "err", err)
-	}
-
-	if m.GetValue() == "" {
-		m.Value = "bonne"
-	}
-
-	req := ctx.CreateRequest(roc.NewIdentifier("res://name"))
-
-	req.SetArgumentByValue("nameArg", roc.NewRepresentation(m))
-
-	nameResp, err := ctx.IssueRequest(req)
+	// issue subrequest to upper case the name
+	upperID := roc.NewIdentifier("res://toUpper")
+	req := ctx.CreateRequest(upperID)
+	req.SetArgumentByValue("value", name)
+	upped, err := ctx.IssueRequest(req)
 	if err != nil {
 		log.Error("failed to dispatch subrequest request", "error", err)
 	}
 
-	r := new(proto.String)
-	err = nameResp.MarshalTo(r)
-	if err != nil {
-		r.Value = "MISSING NAME"
+	s := new(proto.String)
+	if upped.Is(s) {
+		err := upped.MarshalTo(s)
+		if err != nil {
+			return err
+		}
 	}
 
-	r.Value = fmt.Sprintf("hello world: %s", r.Value)
+	return fmt.Sprintf("hello world: %s\n", s.Value)
 
-	return r
 }
 
 func main() {
-	log.Error("GREETER MAIN")
 	endpoint := New()
 	roc.Serve(endpoint)
